FIRMWARE - Added 500 Hz sampling rate for Accelerometer / Gyroscope - Corrected bug that prevented Mass Storage from working properly after first connection - Updated STM32 FS USB libraries SOFTWARE - Added support for 500Hz sampling rate in the WIMU Studio for new firmware

+++ Firmware/WIMU3/inc/ble.h
@@ -115,7 +115,7 @@ typedef struct
 
 typedef struct {
   uint16_t  enabled_modules;
-  uint8_t   sampling_rate;
+  uint16_t  sampling_rate;
   uint8_t   acc_range;
   uint8_t   gyro_range;
   uint8_t   mag_range;
